Encorporated Login/Signup
I have structured the website so that you cannot access any page unless you sign up or login. (php is not working so there is a temporary anchor tag to demo how it will work) when you log in, you are directed to the normal index page with the whole nav bar.